Move generic types/functions to core
