+++ builtin/seq.go
@@ -1,20 +1,17 @@
 package builtin
 
 import (
-	"fmt"
-	"strings"
-
 	"github.com/spy16/slurp/core"
 )
 
 var (
 	_ core.Any = (*LinkedList)(nil)
-	_ Seq      = (*LinkedList)(nil)
+	_ core.Seq = (*LinkedList)(nil)
 )
 
 // Cons returns a new seq with `v` added as the first and `seq` as the rest.
 // seq can be nil as well.
-func Cons(v core.Any, seq Seq) (Seq, error) {
+func Cons(v core.Any, seq core.Seq) (core.Seq, error) {
 	newSeq := &LinkedList{
 		first: v,
 		rest:  seq,
@@ -32,64 +29,14 @@ func Cons(v core.Any, seq Seq) (Seq, error) {
 	return newSeq, nil
 }
 
-// SeqString returns a string representation for the sequence with given prefix
-// suffix and separator.
-func SeqString(seq Seq, begin, end, sep string) (string, error) {
-	var b strings.Builder
-	b.WriteString(begin)
-	err := ForEach(seq, func(item core.Any) (bool, error) {
-		if sxpr, ok := item.(SExpressable); ok {
-			s, err := sxpr.SExpr()
-			if err != nil {
-				return false, err
-			}
-			b.WriteString(s)
-
-		} else {
-			b.WriteString(fmt.Sprintf("%v", item))
-		}
-
-		b.WriteString(sep)
-		return false, nil
-	})
-
-	if err != nil {
-		return "", err
-	}
-
-	return strings.TrimRight(b.String(), sep) + end, err
-}
-
-// ForEach reads from the sequence and calls the given function for each item.
-// Function can return true to stop the iteration.
-func ForEach(seq Seq, call func(item core.Any) (bool, error)) (err error) {
-	var v core.Any
-	var done bool
-	for seq != nil {
-		if v, err = seq.First(); err != nil || v == nil {
-			break
-		}
-
-		if done, err = call(v); err != nil || done {
-			break
-		}
-
-		if seq, err = seq.Next(); err != nil {
-			break
-		}
-	}
-
-	return
-}
-
 // NewList returns a new linked-list containing given values.
-func NewList(items ...core.Any) Seq {
+func NewList(items ...core.Any) core.Seq {
 	if len(items) == 0 {
-		return Seq((*LinkedList)(nil))
+		return core.Seq((*LinkedList)(nil))
 	}
 
 	var err error
-	lst := Seq(&LinkedList{})
+	lst := core.Seq(&LinkedList{})
 	for i := len(items) - 1; i >= 0; i-- {
 		if lst, err = Cons(items[i], lst); err != nil {
 			panic(err)
@@ -103,7 +50,7 @@ func NewList(items ...core.Any) Seq {
 type LinkedList struct {
 	count int
 	first core.Any
-	rest  Seq
+	rest  core.Seq
 }
 
 // SExpr returns a valid s-expression for LinkedList.
@@ -112,11 +59,11 @@ func (ll *LinkedList) SExpr() (string, error) {
 		return "()", nil
 	}
 
-	return SeqString(ll, "(", ")", " ")
+	return core.SeqString(ll, "(", ")", " ")
 }
 
 // Conj returns a new list with all the items added at the head of the list.
-func (ll *LinkedList) Conj(items ...core.Any) (res Seq, err error) {
+func (ll *LinkedList) Conj(items ...core.Any) (res core.Seq, err error) {
 	if ll == nil {
 		res = &LinkedList{}
 	} else {
@@ -141,7 +88,7 @@ func (ll *LinkedList) First() (core.Any, error) {
 }
 
 // Next returns the tail of the list.
-func (ll *LinkedList) Next() (Seq, error) {
+func (ll *LinkedList) Next() (core.Seq, error) {
 	if ll == nil {
 		return nil, nil
 	}
@@ -155,54 +102,4 @@ func (ll *LinkedList) Count() (int, error) {
 	}
 
 	return ll.count, nil
-}
-
-func seqEq(s1, s2 Seq) (bool, error) {
-	if sEq, ok := s1.(EqualityProvider); ok {
-		return sEq.Equals(s2)
-	} else if sEq, ok := s2.(EqualityProvider); ok {
-		return sEq.Equals(s1)
-	}
-
-	if s1 == nil && s2 == nil {
-		return true, nil
-	} else if (s1 == nil && s2 != nil) ||
-		(s1 != nil && s2 == nil) {
-		return false, nil
-	}
-
-	c1, err := s1.Count()
-	if err != nil {
-		return false, err
-	}
-
-	c2, err := s2.Count()
-	if err != nil {
-		return false, err
-	}
-
-	if c1 != c2 {
-		return false, nil
-	}
-
-	bothEqual := true
-	for i := 0; i < c1; i++ {
-		v1, err := s1.First()
-		if err != nil {
-			return false, err
-		}
-
-		v2, err := s2.First()
-		if err != nil {
-			return false, err
-		}
-
-		eq, err := Eq(v1, v2)
-		if err != nil {
-			return false, err
-		}
-		bothEqual = bothEqual && eq
-	}
-
-	return bothEqual, nil
 }
