+++ slurp.go
@@ -9,9 +9,9 @@ import (
 )
 
 // New returns a new slurp interpreter session.
-func New(opts ...Option) *Instance {
+func New(opts ...Option) *Interpreter {
 	buf := bytes.Buffer{}
-	ins := &Instance{
+	ins := &Interpreter{
 		buf:    &buf,
 		reader: reader.New(&buf),
 	}
@@ -25,10 +25,10 @@ func New(opts ...Option) *Instance {
 
 // Option values can be used with New() to customise slurp instance
 // during initialisation.
-type Option func(ins *Instance)
+type Option func(ins *Interpreter)
 
-// Instance represents a Slurp interpreter session.
-type Instance struct {
+// Interpreter represents a Slurp interpreter session.
+type Interpreter struct {
 	env      core.Env
 	buf      *bytes.Buffer
 	reader   *reader.Reader
@@ -36,14 +36,14 @@ type Instance struct {
 }
 
 // Eval performs syntax analysis of the given form to produce an Expr
-// and evalautes the Expr for result.
-func (ins *Instance) Eval(form core.Any) (core.Any, error) {
-	return builtin.Eval(ins.env, ins.analyzer, form)
+// and evaluates the Expr for result.
+func (ins *Interpreter) Eval(form core.Any) (core.Any, error) {
+	return core.Eval(ins.env, ins.analyzer, form)
 }
 
-// EvalStr reads forms from the given straing and evalautes it for
+// EvalStr reads forms from the given string and evaluates it for
 // result.
-func (ins *Instance) EvalStr(s string) (core.Any, error) {
+func (ins *Interpreter) EvalStr(s string) (core.Any, error) {
 	if _, err := ins.buf.WriteString(s); err != nil {
 		return nil, err
 	}
@@ -63,7 +63,7 @@ func (ins *Instance) EvalStr(s string) (core.Any, error) {
 
 // Bind can be used to set global bindings that will be available while
 // executing forms.
-func (ins *Instance) Bind(vals map[string]core.Any) error {
+func (ins *Interpreter) Bind(vals map[string]core.Any) error {
 	for k, v := range vals {
 		if err := ins.env.Bind(k, v); err != nil {
 			return err
@@ -75,9 +75,9 @@ func (ins *Instance) Bind(vals map[string]core.Any) error {
 // WithEnv sets the environment to be used by the slurp instance. If
 // env is nil, the default map-env will be used.
 func WithEnv(env core.Env) Option {
-	return func(ins *Instance) {
+	return func(ins *Interpreter) {
 		if env == nil {
-			env = builtin.New(nil)
+			env = core.New(nil)
 		}
 		ins.env = env
 	}
@@ -87,7 +87,7 @@ func WithEnv(env core.Env) Option {
 // syntax analysis and macro expansions. If nil, uses builtin analyzer
 // with standard special forms.
 func WithAnalyzer(a core.Analyzer) Option {
-	return func(ins *Instance) {
+	return func(ins *Interpreter) {
 		if a == nil {
 			a = &builtin.Analyzer{
 				Specials: map[string]builtin.ParseSpecial{
